refactor(courses): Migrate shop page to /courses with design system colors
Remove /shop route entirely and replace /courses page with shop design
using real course data from database. Use only design system colors.

Color System Changes:
- Remove 10 redundant shop-specific color variables
- Keep only --badge-orange and --badge-orange-bg (unique colors)
- Replace --shop-primary → --primary throughout
- Replace --shop-success → --success throughout
- Replace --shop-warning → --warning throughout
- Replace badge colors with existing design system colors

Component Changes:
- CourseCard: New design using real Course data, preserves
  localStorage update + navigate to / behavior
- PlanCard: Moved from shop to courses, updated to use design system
  colors (--primary, --success, --warning)
- MembershipPlans: New client component with hardcoded plans data
- CoursesPage: Server component with queryPublishedCourses(),
  membership plans section, and shop-inspired layout

Files Removed:
- src/app/(frontend)/shop/ directory (entire route)

All components now use only generic design system colors as requested.
No feature-specific colors remain. Text is hardcoded in Hebrew to
maintain server component pattern.
