+++ src/app/(frontend)/courses/_components/CourseCard/index.tsx
@@ -1,19 +1,18 @@
 'use client'
 
+import { BookOpen, CheckCircle } from 'lucide-react'
+import { cn } from '@/infra/utils/ui'
 import { useRouterWithLoading } from '@/infra/loading/hooks/useRouterWithLoading'
+import { setUserProfile, getUserProfile } from '@/client/state/localStorage/userProfile'
 import type { Course } from '@/payload-types'
 import { useTranslations } from '@/ui/web/providers/I18n'
-import { Card, CardDescription, CardFooter, CardHeader, CardTitle } from '@/ui/web/components/card'
-import { Button } from '@/ui/web/components/button'
-import { Badge } from '@/ui/web/components/badge'
-import { ArrowRight } from 'lucide-react'
-import { setUserProfile, getUserProfile } from '@/client/state/localStorage/userProfile'
 
 interface CourseCardProps {
   course: Course
+  isOwned?: boolean
 }
 
-export function CourseCard({ course }: CourseCardProps) {
+export function CourseCard({ course, isOwned = false }: CourseCardProps) {
   const t = useTranslations('courses')
   const router = useRouterWithLoading()
 
@@ -45,38 +44,94 @@ export function CourseCard({ course }: CourseCardProps) {
     router.push('/')
   }
 
+  const getButtonClasses = () => {
+    if (isOwned) {
+      return 'bg-[hsl(var(--success))]/10 text-[hsl(var(--success))] px-6 py-2.5 rounded-xl'
+    }
+    return 'bg-muted px-6 py-2.5 rounded-xl hover:bg-[hsl(var(--primary-soft))] transition-colors'
+  }
+
+  const getButtonText = () => {
+    if (isOwned) {
+      return t('openCourse')
+    }
+    return t('openCourse')
+  }
+
+  const borderClass = isOwned
+    ? 'border-2 border-[hsl(var(--primary))]/20'
+    : 'border border-transparent hover:border-[hsl(var(--primary-soft))]'
+
   return (
-    <div className="group relative">
-      <div className="absolute -inset-0.5 bg-gradient-to-r from-primary/50 via-accent/50 to-primary/50 rounded-xl opacity-0 group-hover:opacity-100 blur transition duration-500" />
-      <Card className="relative overflow-hidden border-border bg-card backdrop-blur-sm hover:bg-card transition-all duration-300 h-full flex flex-col shadow-xl">
-        <div className="absolute inset-0 bg-gradient-to-br from-primary/15 via-transparent to-accent/15 opacity-0 group-hover:opacity-100 transition-opacity duration-500" />
+    <button
+      onClick={handleCourseSelect}
+      className={cn(
+        'relative bg-card p-6 rounded-[2rem] flex flex-col text-right w-full',
+        borderClass,
+        'shadow-[0_1px_2px_0_rgba(60,64,67,.3),0_1px_3px_1px_rgba(60,64,67,.15)]',
+        'transition-all active:scale-[0.98] hover:-translate-y-0.5',
+      )}
+    >
+      {isOwned && (
+        <span
+          className="absolute -top-3 left-6 bg-[hsl(var(--success))] text-white px-4 py-1 rounded-full shadow-md uppercase tracking-wider"
+          style={{ fontSize: '9px', fontWeight: 900 }}
+        >
+          הקורס שלך
+        </span>
+      )}
 
-        <CardHeader className="relative pb-4">
+      <div className="mb-6 flex justify-between items-start gap-4">
+        <div className="flex-1">
           {course.courseLabel && (
-            <Badge variant="secondary" className="w-fit mb-3 text-xs font-medium">
+            <span
+              className="block mb-1 uppercase tracking-widest text-primary"
+              style={{ fontSize: '10px', fontWeight: 900 }}
+            >
               {course.courseLabel}
-            </Badge>
+            </span>
           )}
-          <CardTitle className="text-2xl font-bold bg-gradient-to-br from-foreground to-foreground/70 bg-clip-text">
-            {course.title}
-          </CardTitle>
-          {course.description && (
-            <CardDescription className="line-clamp-3 text-muted-foreground/80 mt-2">
-              {course.description}
-            </CardDescription>
-          )}
-        </CardHeader>
-
-        <CardFooter className="relative mt-auto pt-4">
-          <Button
-            onClick={handleCourseSelect}
-            className="w-full group/btn bg-primary text-primary-foreground hover:bg-primary/90 shadow-lg shadow-primary/20 hover:shadow-primary/30 transition-all duration-300 flex items-center justify-center gap-2"
+          <h4
+            className="text-card-foreground text-right"
+            style={{ fontSize: '20px', fontWeight: 900 }}
           >
-            {t('openCourse')}
-            <ArrowRight className="w-4 h-4 group-hover/btn:translate-x-1 transition-transform" />
-          </Button>
-        </CardFooter>
-      </Card>
-    </div>
+            {course.title}
+          </h4>
+          {course.description && (
+            <p
+              className="text-muted-foreground mt-1 line-clamp-2 text-right"
+              style={{ fontSize: '12px' }}
+            >
+              {course.description}
+            </p>
+          )}
+        </div>
+        <div
+          className={cn(
+            'w-12 h-12 rounded-2xl flex items-center justify-center flex-shrink-0',
+            isOwned ? 'bg-[hsl(var(--success))]/10' : 'bg-muted',
+          )}
+        >
+          {isOwned ? (
+            <CheckCircle className="w-6 h-6 text-[hsl(var(--success))]" />
+          ) : (
+            <BookOpen className="w-6 h-6 text-primary" />
+          )}
+        </div>
+      </div>
+
+      <div className="mt-auto flex items-center justify-between pt-6 border-t border-border">
+        {/* Price placeholder - could be added when pricing is implemented */}
+        <span className="text-card-foreground" style={{ fontSize: '20px', fontWeight: 900 }}>
+          {/* ₪149 */}
+        </span>
+        <span
+          className={cn(getButtonClasses(), 'text-primary hover:text-[hsl(var(--primary))]')}
+          style={{ fontSize: '12px', fontWeight: 700 }}
+        >
+          {getButtonText()}
+        </span>
+      </div>
+    </button>
   )
 }
